requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
arxiv
PyPDF2
yt_dlp
requests
EOF
echo "requirements.txt created."

# Install dependencies
pip install -r requirements.txt
echo "Dependencies installed."

# Create .env file with placeholders
cat <<EOL > .env
OPENAI_API_KEY=your_openai_api_key
GITHUB_API_KEY=your_github_api_key
GROQ_API_KEY=your_groq_api_key
EOF
echo ".env file created with placeholders."

# Create README.md
cat <<EOL > README.md
# Arxiv Podcast Generator

This project is a FastAPI backend that accepts an arXiv URL, processes the paper using OpenAI's GPT-4o model to generate and refine a podcast plan, and ultimately produces a podcast script. The process involves multiple API calls to OpenAI to plan, critique, and regenerate the podcast content to ensure it is engaging and understandable for a diverse audience.

## Project Structure

```
arxiv_podcast_generator/
├── app/
│   └── main.py
├── arxiv_papers/
├── downloads/
├── scripts/
│   └── setup_project.sh
├── tests/
├── .env
├── .gitignore
├── README.md
├── requirements.txt
└── venv/
```

## Setup Instructions

1. **Clone the Repository**

    ```bash
    git clone https://github.com/yourusername/arxiv_podcast_generator.git
    cd arxiv_podcast_generator
    ```

2. **Run the Setup Script**

    Make sure the setup script is executable:

    ```bash
    chmod +x scripts/setup_project.sh
    ./scripts/setup_project.sh
    ```

3. **Configure Environment Variables**

    Update the `.env` file with your OpenAI API key and other necessary credentials.

4. **Run the FastAPI Server**

    ```bash
    uvicorn app.main:app --reload
    ```

    The API will be available at `http://localhost:8000`.

## API Endpoint

- **Create Podcast**

    **URL**: `/create-podcast/`

    **Method**: `POST`

    **Payload**:

    ```json
    {
      "url": "https://arxiv.org/abs/2106.14834"
    }
    ```

    **Response**:

    ```json
    {
      "podcast_plan": { ... },
      "podcast_script": { ... },
      "critique": "..."
    }
    ```

## Dependencies

- Python 3.7+
- FastAPI
- Uvicorn
- Pydantic
- OpenAI
- python-dotenv
- arxiv
- PyPDF2
- yt_dlp
- requests

## License

This project is licensed under the MIT License.
